Scenario API

+++ reservoir_management/scenario/views.py
@@ -37,8 +37,8 @@ def get_data(request, district_id, year):
 
 def get_simulator(request):
     # Extract query parameters
-    evaporation = int(request.GET.get("evaporation", "0")) 
-    rainfall = int(request.GET.get("rainfall", "0"))
+    evaporation = float(request.GET.get("evaporation", "0")) 
+    rainfall = float(request.GET.get("rainfall", "0"))
     population = int(request.GET.get("population", "0"))
     district_id = int(request.GET.get("district_id","0"))
     print(district_id)
